requirements.txt
# Core dependencies
click>=8.1.7,<8.2.0
rich>=13.7.0,<13.8.0
pyyaml>=6.0.1,<6.1.0
pytest>=8.0.2
pytest-asyncio>=0.23.5

# Serverless dependencies
mangum>=0.17.0
aws-lambda-powertools>=2.30.0
serverless-wsgi>=3.0.0

# LLM Providers
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
cohere==4.50
huggingface-hub>=0.21.3

# Optional dependencies
plotly>=5.19.0
pandas>=2.2.1
aiohttp>=3.9.3
prometheus-client>=0.20.0

# UI dependencies
streamlit>=1.31.1
uvicorn>=0.27.1,<0.28.0
jinja2>=3.1.3

# Core Web UI dependencies
fastapi[all]>=0.110.0,<0.111.0
python-multipart>=0.0.9
httpx>=0.26.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
